requirements.txt
alpaca==1.0.0
alpaca_trade_api==3.0.0
arch==5.4.0
Flask==2.2.3
hyperopt==0.2.7
kneed==0.8.5
matplotlib==3.7.1
numpy==1.24.2
pandas==1.5.3
pandas_datareader==0.10.0
plotly==5.14.1
pykalman==0.9.5
python-dotenv==1.0.0
Requests==2.31.0
scikit_learn==1.2.1
statsmodels==0.13.5
ta==0.10.3
yfinance==0.2.18
